+++ quickSort.h
@@ -1,6 +1,6 @@
 #ifndef QUICKSORT_H
 #define QUICKSORT_H
 
-void quickSort(int array[], int low, int high, int &counter);
+void quickSort(int array[], int low, int high, long long &counter);
 
 #endif
